refactor(autonomous): extract GitWorkspaceService verbatim from orchestrator (#2044 Phase B T5)
#2041/#2042/#2043 behaviour unchanged; orchestrator keeps delegating wrappers.

Refs #2044
